+++ Captain America and The Avengers/Framework/dxutil.cpp
@@ -145,7 +145,7 @@ HRESULT DXUtil_WriteStringRegKey( HKEY hKey, TCHAR* strRegName,
 {
     if( ERROR_SUCCESS != RegSetValueEx( hKey, strRegName, 0, REG_SZ, 
                                         (BYTE*)strValue, 
-                                        (_tcslen(strValue)+1)*sizeof(TCHAR) ) )
+                                        (_tcslen(strValue) + 1 )*sizeof(TCHAR) ) )
         return E_FAIL;
 
     return S_OK;
@@ -468,7 +468,7 @@ VOID DXUtil_ConvertAnsiStringToWide( WCHAR* wstrDestination, const CHAR* strSour
         return;
 
     if( cchDestChar == -1 )
-        cchDestChar = strlen(strSource)+1;
+        cchDestChar = (int)(strlen(strSource) + 1);
 
     MultiByteToWideChar( CP_ACP, 0, strSource, -1, 
                          wstrDestination, cchDestChar-1 );
@@ -491,8 +491,8 @@ VOID DXUtil_ConvertWideStringToAnsi( CHAR* strDestination, const WCHAR* wstrSour
     if( strDestination==NULL || wstrSource==NULL )
         return;
 
-    if( cchDestChar == -1 )
-        cchDestChar = wcslen(wstrSource)+1;
+	if (cchDestChar == -1)
+		cchDestChar = (int)(wcslen(wstrSource) + 1);
 
     WideCharToMultiByte( CP_ACP, 0, wstrSource, -1, strDestination, 
                          cchDestChar-1, NULL, NULL );
