+++ Captain America and The Avengers/PittsburghBossMap.cpp
@@ -23,13 +23,10 @@ void PittsburghBossMap::Draw(D3DCOLOR ColorMode)
 	LPD3DXSPRITE Hander = d3d::GetInstance()->GetSpriteHander();
 	D3DXMATRIX OldMatrix;
 	Hander->GetTransform(&OldMatrix);
-
 	D3DXVECTOR2 pos = Camera::GetInstance()->GetCameraPosition();
 	int gStartX = (int)(pos.x / PIXEL_PER_TITLE);
 	int gStartY = Height - (int)(pos.y / PIXEL_PER_TITLE) - GRID_HIGHT_BLOCK - 1;
-
 	Hander->Begin(D3DXSPRITE_ALPHABLEND);
-
 	int x = (PIXEL_PER_TITLE >> 1) - ((int)pos.x % PIXEL_PER_TITLE);
 	for (int i = gStartX; i <= gStartX + GRID_WIDTH_BLOCK; i++)
 	{
@@ -44,11 +41,11 @@ void PittsburghBossMap::Draw(D3DCOLOR ColorMode)
 			switch (Map[j][i])
 			{
 			case 0:
-				CaroDown->SetPosition(x, y);
+				CaroDown->SetPosition((FLOAT)x, (FLOAT)y);
 				CaroDown->ImperiouslyDraw();
 				break;
 			case 1:
-				CaroUp->SetPosition(x, y);
+				CaroUp->SetPosition((FLOAT)x, (FLOAT)y);
 				CaroUp->ImperiouslyDraw();
 				break;
 			default:
@@ -58,7 +55,6 @@ void PittsburghBossMap::Draw(D3DCOLOR ColorMode)
 		}
 		x += PIXEL_PER_TITLE;
 	}
-
 	Hander->End();
 	Hander->SetTransform(&OldMatrix);
 }
